Feat/docker

+++ test/auth.e2e-spec.ts
@@ -29,7 +29,10 @@ describe('Authentication (e2e)', () => {
   });
 
   afterAll(async () => {
-    await dataSource.dropDatabase();
+    if (dataSource && dataSource.isInitialized) {
+      await dataSource.dropDatabase();
+      await dataSource.destroy();
+    }
     await app.close();
   });
